Improve naming
